TE: Tests refactoring
