+++ tests/E2E/test/campaigns/high/01_orders/6_check_FO_order_created_in_BO.js
@@ -10,7 +10,7 @@ scenario('Check order created in BO', client => {
     test('should go to "Orders" page', () => client.goToSubtabMenuPage(OrderPage.orders_subtab, OrderPage.order_submenu));
     test('should search the order created by reference', () => client.waitAndSetValue(OrderPage.reference_search, global.tab['reference']));
     test('should go to search order', () => client.waitForExistAndClick(OrderPage.search_order_button));
-    test('should go to the order ', () => client.waitForExistAndClick(OrderPage.view_order_button));
+    test('should go to the order ', () => client.scrollWaitForExistAndClick(OrderPage.view_order_button));
     test('should check the customer name ', () => client.checkTextValue(OrderPage.customer_name, 'John DOE', 'contain'));
     test('should status be equal to Awaiting bank wire payment ', () => client.checkTextValue(OrderPage.order_check_status, 'Awaiting bank wire payment', "equal"));
     test('should check the shipping price', () => client.checkTextValue(OrderPage.check_shipping_price, global.tab['shipping_price'], "equal"));
