+++ tests/E2E/test/campaigns/high/02_product/5_check_default_product_sort.js
