+++ tests/E2E/test/campaigns/high/03_category/2_update_category.js
