+++ tests/E2E/test/campaigns/high/03_category/3_delete_category.js
