+++ tests/E2E/test/campaigns/high/04_attribute/1_create_attribute.js
@@ -2,43 +2,41 @@ const {AccessPageBO} = require('../../../selectors/BO/access_page');
 const {AccessPageFO} = require('../../../selectors/FO/access_page');
 const {CatalogPage} = require('../../../selectors/BO/catalogpage/index');
 const {AttributeSubMenu} = require('../../../selectors/BO/catalogpage/attribute_submenu');
-const {FeatureSubMenu} = require('../../../selectors/BO/catalogpage/feature_submenu');
 const {AddProductPage} = require('../../../selectors/BO/add_product_page');
 const {SearchProductPage} = require('../../../selectors/FO/search_product_page');
-const common_scenarios = require('../2_product/product');
+const common_scenarios = require('../02_product/product');
 
-productData = {
-  name: 'Feature',
-  quantity: "10",
-  price: '5',
-  image_name: 'image_test.jpg',
-  feature:{
-    name: 'Feature',
-    value: 'feature value'
-  }
+var productData = {
+    name: 'Att',
+    quantity: "10",
+    price: '5',
+    image_name: 'image_test.jpg',
+    attribute:{
+      name: 'attribute' ,
+      variation_quantity: '10'
+    }
 };
 
-scenario('Create "Feature"', () => {
+scenario('Create "Attribute"', () => {
   scenario('Login in the Back Office', client => {
     test('should open the browser', () => client.open());
     test('should login successfully in the Back Office', () => client.signInBO(AccessPageBO));
   }, 'attribute_and_feature');
-  scenario('Create a new "Feature"', client => {
+  scenario('Create a new "Attribute"', client => {
     test('Should go to "Attributes & Features" page', () => client.goToSubtabMenuPage(CatalogPage.menu_button, AttributeSubMenu.submenu));
-    test('should click on "Feature" subtab', () => client.waitForExistAndClick(FeatureSubMenu.tabmenu));
-    test('should click on "Add new feature" button', () => client.waitForExistAndClick(FeatureSubMenu.add_new_feature));
-    test('should set the "Name" input', () => client.waitAndSetValue(FeatureSubMenu.name_input, 'Feature' + date_time));
-    test('should click on "Save" button', () => client.waitForExistAndClick(FeatureSubMenu.save_button));
+    test('should click on "Add new attribute" button', () => client.waitForExistAndClick(AttributeSubMenu.add_new_attribute));
+    test('should set the "Name" input', () => client.waitAndSetValue(AttributeSubMenu.name_input, 'attribute' + date_time));
+    test('should set the "Public name" input', () => client.waitAndSetValue(AttributeSubMenu.public_name_input, 'attribute' + date_time));
+    test('should choose the "Type" of attribute', () => client.waitAndSelectByValue(AttributeSubMenu.type_select, 'radio'));
+    test('should click on "Save" button', () => client.waitForExistAndClick(AttributeSubMenu.save_button));
     test('should verify the appearance of the green validation', () => client.checkTextValue(CatalogPage.success_panel, '×\nSuccessful creation.'));
-    test('should search for the created feature', () => client.searchByValue(FeatureSubMenu.search_input, FeatureSubMenu.search_button, 'Feature' + date_time));
-    test('should select the created feature', () => client.waitForExistAndClick(FeatureSubMenu.selected_feature));
-    test('should click on "Add new feature value" button', () => client.waitForExistAndClick(FeatureSubMenu.add_value_button));
-    test('should set the "Value" input', () => client.waitAndSetValue(FeatureSubMenu.value_input, 'feature value'));
-    test('should click on "Save" button', () => client.waitForExistAndClick(FeatureSubMenu.save_value_button));
+    test('should search for the created attribute', () => client.searchByValue(AttributeSubMenu.search_input, AttributeSubMenu.search_button, 'attribute' + date_time));
+    test('should select the created attribute', () => client.waitForExistAndClick(AttributeSubMenu.selected_attribute));
+    test('should add value to the created attribute', () => client.addValueToAttribute(AttributeSubMenu));
     test('should verify the appearance of the green validation', () => client.checkTextValue(CatalogPage.success_panel, '×\nSuccessful creation.'));
   }, 'attribute_and_feature');
   scenario('Logout from the Back Office', client => {
-    test('should logout successfully from the Back Office', () => client.signOutBO());
+    test('should logout successfully from Back Office', () => client.signOutBO());
   }, 'attribute_and_feature');
 }, 'attribute_and_feature', true);
 
@@ -53,17 +51,19 @@ scenario('Create "Product"', () => {
   }, 'attribute_and_feature');
 }, 'attribute_and_feature', true);
 
-scenario('Check the feature creation', () => {
+scenario('Check the attribute creation', () => {
   scenario('Login in the Front Office', client => {
     test('should open the browser', () => client.open());
     test('should login successfully in the Front Office', () => client.signInFO(AccessPageFO));
   }, 'attribute_and_feature');
-  scenario('Check that the feature is well created in Front Office', client => {
+  scenario('Check that the attribute is well created in the Front Office', client => {
     test('should set the shop language to "English"', () => client.changeLanguage('english'));
-    test('should search for the product', () => client.searchByValue(SearchProductPage.search_input, SearchProductPage.search_button, 'Feature'));
+    test('should search for the product', () => client.searchByValue(SearchProductPage.search_input, SearchProductPage.search_button, productData.name + date_time));
     test('should go to the product page', () => client.waitForExistAndClick(SearchProductPage.product_result_name));
-    test('should check that the name of created feature is : "'+ 'Feature' + date_time +'"', () => client.checkTextValue(SearchProductPage.feature_name, 'Feature' + date_time));
-    test('should check that the value of created feature is : "Feature Value"', () => client.checkTextValue(SearchProductPage.feature_value, 'Feature Value'));
+    test('should check that the product attribute name is "'+'attribute' + date_time+'"', () => client.checkTextValue(SearchProductPage.attribute_name, 'attribute' + date_time));
+    test('should check that the first attribute value is equal to 10', () => client.checkTextValue(SearchProductPage.attribute_value_1, '10'));
+    test('should check that the second attribute value is equal to 20', () => client.checkTextValue(SearchProductPage.attribute_value_2, '20'));
+    test('should check that the third attribute value is equal to 30', () => client.checkTextValue(SearchProductPage.attribute_value_3, '30'));
   }, 'attribute_and_feature');
   scenario('Logout from the Front Office', client => {
     test('should logout successfully from the Front Office', () => client.signOutFO(AccessPageFO));
