+++ tests/E2E/test/campaigns/high/04_attribute/2_update_attribute.js
@@ -34,12 +34,12 @@ scenario('Check the attribute modification', () => {
   }, 'attribute_and_feature');
   scenario('Check that the attribute is well updated in the Front Office', client => {
     test('should set the shop language to "English"', () => client.changeLanguage('english'));
-    test('should search for the product', () => client.searchByValue(SearchProductPage.search_input, SearchProductPage.search_button, 'attribute'));
+    test('should search for the product', () => client.searchByValue(SearchProductPage.search_input, SearchProductPage.search_button, 'Att' + date_time));
     test('should go to the product page', () => client.waitForExistAndClick(SearchProductPage.product_result_name));
-    test('should check that the product attribute name is "'+'attribute' + date_time+'update"', () => client.checkTextValue(SearchProductPage.attribut_name, 'attribute' + date_time + 'update'));
-    test('should check that the first attribute value is equal to 40', () => client.checkTextValue(SearchProductPage.attribut_value_1, '40'));
-    test('should check that the second attribute value is equal to 20', () => client.checkTextValue(SearchProductPage.attribut_value_2, '20'));
-    test('should check that the third attribute value is equal to 30', () => client.checkTextValue(SearchProductPage.attribut_value_3, '30'));
+    test('should check that the product attribute name is "'+'attribute' + date_time+'update"', () => client.checkTextValue(SearchProductPage.attribute_name, 'attribute' + date_time + 'update'));
+    test('should check that the first attribute value is equal to 40', () => client.checkTextValue(SearchProductPage.attribute_value_1, '40'));
+    test('should check that the second attribute value is equal to 20', () => client.checkTextValue(SearchProductPage.attribute_value_2, '20'));
+    test('should check that the third attribute value is equal to 30', () => client.checkTextValue(SearchProductPage.attribute_value_3, '30'));
   }, 'attribute_and_feature');
   scenario('Logout from the Front Office', client => {
     test('should logout successfully from the Front Office', () => client.signOutFO(AccessPageFO));
