+++ tests/E2E/test/campaigns/high/06_catalogbulkaction/catalogbulkaction.js
