+++ tests/E2E/test/campaigns/high/07_manufacturer/create_manufacturers.js
