+++ tests/E2E/test/campaigns/high/08_stocks/2_single_product_quantity.js
