+++ tests/E2E/test/campaigns/high/09_customer/1_create_customer.js
