+++ tests/E2E/test/campaigns/high/11_international/translation.js
@@ -3,7 +3,7 @@ const {AccessPageFO} = require('../../../selectors/FO/access_page');
 const {Translations} = require('../../../selectors/BO/international/translations');
 const {AddProductPage} = require('../../../selectors/BO/add_product_page');
 
-scenario('Edit a translation', client => {
+scenario('Edit a translation', () => {
   scenario('Open the browser and connect to the BO', client => {
     test('should open the browser', () => client.open());
     test('should log in successfully in BO', () => client.signInBO(AccessPageBO));
@@ -27,10 +27,10 @@ scenario('Edit a translation', client => {
     test('should login successfully in the Front Office', () => client.signInFO(AccessPageFO));
   }, 'common_client');
   scenario('Check the change of "Sign in" to "Sign in English" ', client => {
-    test('should set the shop language to "English"', () => client.languageChange('english'));
+    test('should set the shop language to "English"', () => client.changeLanguage('english'));
     test('should check the "Sign in" button text ', () => client.checkTextValue(Translations.sign_out_FO_text, 'Sign out English', "contain"));
   }, 'common_client');
   scenario('Logout from the Front Office', client => {
     test('should logout successfully from the Front Office', () => client.signOutFO(AccessPageFO));
   }, 'attribute_and_feature');
-}, 'common_client');
+}, 'common_client', true);
