+++ tests/E2E/test/selectors/BO/index.js
@@ -8,7 +8,7 @@ module.exports = Object.assign({},
   require('./order_page'),
   require('./create_order'),
   require('./product_list'),
-  require('./shopParameters')
+  require('./shopParameters'),
   require('./international')
 );
