+++ tests/E2E/test/selectors/BO/order_page.js
@@ -12,7 +12,7 @@ module.exports = {
     order_quantity: '//*[@id="orderProducts"]/tbody/tr[1]/td[4]/span[1]',
     order_check_status: '//*[@id="status"]/div/table/tbody/tr[1]/td[2]',
     check_shipping_Cost: '//*[@id="shipping_table"]/tbody/tr/td[5]/span',
-    check_message_order: '//*[@id="content"]//p[contains(@class, "message-item-text")]',
+    check_message_order: '//*[@id="content"]//div[@class="message-body"]//p[@class="message-item-text"]',
     check_payment_type: '//*[@id="formAddPayment"]/div/table/tbody/tr[1]/td[2]',
     check_quantity: '//*[@id="orderProducts"]/tbody/tr[1]/td[4]/span[1]',
     quantity_value:'//*[@id="orderProducts"]/tbody/tr[1]/td[4]/span[2]/input',
@@ -21,7 +21,7 @@ module.exports = {
     edit_product_button: '//*[@id="orderProducts"]/tbody/tr[1]/td[13]/div/button[1]',
     second_edit_product_button:'//*[@id="orderProducts"]/tbody/tr[1]/td[12]/div/button[1]',
     product_basic_price: '//*[@id="orderProducts"]/tbody/tr[1]/td[3]/div/div/div[1]/div/input',
-    customer_name: '//*[@id="content"]//h2[contains(@class, "page-title")]',
+    customer_name: '//*[@id="content"]//div[@class="message-body"]//h4[@class="message-item-heading"]',
     update_order_status_button: '//*[@id="status"]/form/div/div[2]/button',
     order_submenu: '//*[@id="subtab-AdminOrders"]/a',
     document_submenu:'//*[@id="tabOrder"]/li[2]/a',
