+++ tests/E2E/test/selectors/FO/search_product_page.js
@@ -7,11 +7,11 @@ module.exports = {
     product_result_name: '.h3.product-title > a',
     second_product_result_name: '//*[@id="js-product-list"]/div[1]/article[2]/div/div[1]/h1/a',
     product_result_price: '[itemprop="price"]',
-    attribut_name: '//*[@id="add-to-cart-or-refresh"]/div[1]/div/span',
+    attribute_name: '//*[@id="add-to-cart-or-refresh"]/div[1]/div/span',
     feature_name: '//*[@id="product-details"]/section/dl/dt',
     feature_value: '//*[@id="product-details"]/section/dl/dd',
-    attribut_value_1: '//*[@id="add-to-cart-or-refresh"]/div[1]/div/ul/li[1]/label/span',
-    attribut_value_2: '//*[@id="add-to-cart-or-refresh"]/div[1]/div/ul/li[2]/label/span',
-    attribut_value_3: '//*[@id="add-to-cart-or-refresh"]/div[1]/div/ul/li[3]/label/span',
+    attribute_value_1: '//*[@id="add-to-cart-or-refresh"]/div[1]/div/ul/li[1]/label/span',
+    attribute_value_2: '//*[@id="add-to-cart-or-refresh"]/div[1]/div/ul/li[2]/label/span',
+    attribute_value_3: '//*[@id="add-to-cart-or-refresh"]/div[1]/div/ul/li[3]/label/span',
   }
 };
